requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
APScheduler==3.2.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
blinker==1.4
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
cloud-init==20.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
croniter==0.3.37
cryptography==3.4.7
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dpcpp-cpp-rt==2021.4.0
entrypoints==0.3
fake-useragent==0.1.11
fastapi==0.68.1
Flask==1.0
graia-application-mirai==0.18.4
graia-broadcast==0.8.11
graia-scheduler==0.0.4
gunicorn==19.9.0
h11==0.12.0
httpcore==0.13.7
httplib2==0.14.0
httpx==0.20.0
hyperlink==19.0.0
idna==3.2
importlib-metadata==1.5.0
incremental==16.10.1
intel-cmplr-lib-rt==2021.4.0
intel-cmplr-lic-rt==2021.4.0
intel-opencl-rt==2021.4.0
intel-openmp==2021.4.0
isc==2.0
iterwrapper==0.1.4
itsdangerous==2.0.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.6.3
MarkupSafe==1.1.0
mkl==2021.4.0
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==4.2.0
multidict==5.2.0
natsort==7.1.1
netifaces==0.10.4
numpy==1.20.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.54
pexpect==4.6.0
Pillow==8.3.1
ply==3.11
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycryptodome==3.10.1
pycurl==7.43.0.2
pydantic==1.7.1
PyExecJS==1.5.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.5.3
regex==2020.11.13
requests==2.26.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sos==4.1
soupsieve==2.2.1
ssh-import-id==5.10
starlette==0.14.2
systemd-python==234
tbb==2021.4.0
Twisted==18.9.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
uvicorn==0.15.0
wadllib==1.3.3
Werkzeug==0.15.5
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.7.0
zipp==1.0.0
zope.interface==4.7.1
